+++ examples/data-sources/hrobot_failover/data-source.tf
@@ -1,0 +1,7 @@
+data "hrobot_failover" "example" {
+  ip = "203.0.113.5"
+}
+
+output "failover_routed_to" {
+  value = data.hrobot_failover.example.active_server_ip
+}