+++ examples/data-sources/hrobot_storagebox/data-source.tf
@@ -1,0 +1,3 @@
+data "hrobot_storagebox" "example" {
+  id = 12345
+}