+++ examples/data-sources/hrobot_storageboxes/data-source.tf
@@ -1,0 +1,2 @@
+# All Storage Boxes on the account.
+data "hrobot_storageboxes" "all" {}