+++ examples/data-sources/hrobot_subnets/data-source.tf
@@ -1,0 +1,2 @@
+# All subnets on the account.
+data "hrobot_subnets" "all" {}